+++ Sources/EV3/EV3DrawTextCommand.swift
@@ -32,7 +32,7 @@ public struct EV3DrawTextCommand: EV3DirectCommand {
 		return EV3GenericResponse.self
 	}
 
-	public func payloadDataWithGlobalOffset(offset: UInt8) -> NSData {
+	public func payloadDataWithGlobalOffset(offset: UInt16) -> NSData {
 		let mutableData = NSMutableData()
 		mutableData.appendUInt8(EV3OpCode.UIDraw.rawValue)
 		mutableData.appendUInt8(EV3UIDrawOpSubcode.SelectFont.rawValue)
