+++ Sources/EV3/EV3EnableToplineCommand.swift
@@ -18,7 +18,7 @@ public struct EV3EnableToplineCommand: EV3DirectCommand {
 		return EV3GenericResponse.self
 	}
 
-	public func payloadDataWithGlobalOffset(offset: UInt8) -> NSData {
+	public func payloadDataWithGlobalOffset(offset: UInt16) -> NSData {
 		let mutableData = NSMutableData()
 		mutableData.appendUInt8(EV3OpCode.UIDraw.rawValue)
 		mutableData.appendUInt8(EV3UIDrawOpSubcode.Topline.rawValue)
