+++ Sources/EV3/EV3PlaySoundFileCommand.swift
@@ -20,7 +20,7 @@ public struct EV3PlaySoundFileCommand: EV3DirectCommand {
 		return EV3GenericResponse.self
 	}
 
-	public func payloadDataWithGlobalOffset(offset: UInt8) -> NSData {
+	public func payloadDataWithGlobalOffset(offset: UInt16) -> NSData {
 		let mutableData = NSMutableData()
 		mutableData.appendUInt8(EV3OpCode.Sound.rawValue)
 		mutableData.appendUInt8(EV3SoundOpSubcode.Play.rawValue)
